Move Settings and Unique Ids sync to core
