+++ opensrp-anc/src/main/java/org/smartregister/anc/application/AncSyncConfiguration.java
@@ -1,12 +1,14 @@
-package org.smartregister.anc.service.intent;
+package org.smartregister.anc.application;
 
+import org.smartregister.SyncConfiguration;
 import org.smartregister.anc.BuildConfig;
-import org.smartregister.anc.application.AncApplication;
 import org.smartregister.anc.util.Constants;
 import org.smartregister.repository.AllSharedPreferences;
 
-
-public class SyncIntentService extends org.smartregister.sync.intent.SyncIntentService {
+/**
+ * Created by samuelgithengi on 10/19/18.
+ */
+public class AncSyncConfiguration extends SyncConfiguration {
     @Override
     public int getSyncMaxRetries() {
         return BuildConfig.MAX_SYNC_RETRIES;
@@ -21,5 +23,20 @@ public class SyncIntentService extends org.smartregister.sync.intent.SyncIntentS
     public String getSyncFilterValue() {
         AllSharedPreferences sharedPreferences = AncApplication.getInstance().getContext().userService().getAllSharedPreferences();
         return sharedPreferences.fetchDefaultTeamId(sharedPreferences.fetchRegisteredANM());
+    }
+
+    @Override
+    public int getUniqueIdSource() {
+        return BuildConfig.OPENMRS_UNIQUE_ID_SOURCE;
+    }
+
+    @Override
+    public int getUniqueIdBatchSize() {
+        return BuildConfig.OPENMRS_UNIQUE_ID_BATCH_SIZE;
+    }
+
+    @Override
+    public int getUniqueIdInitialBatchSize() {
+        return BuildConfig.OPENMRS_UNIQUE_ID_INITIAL_BATCH_SIZE;
     }
 }
