+++ opensrp-anc/src/main/java/org/smartregister/anc/fragment/BaseRegisterFragment.java
@@ -32,7 +32,6 @@ import org.smartregister.anc.contract.RegisterFragmentContract;
 import org.smartregister.anc.cursor.AdvancedMatrixCursor;
 import org.smartregister.anc.domain.AttentionFlag;
 import org.smartregister.anc.event.SyncEvent;
-import org.smartregister.anc.job.SyncSettingsServiceJob;
 import org.smartregister.anc.provider.RegisterProvider;
 import org.smartregister.anc.util.Constants;
 import org.smartregister.anc.util.DBConstants;
@@ -44,6 +43,7 @@ import org.smartregister.cursoradapter.RecyclerViewFragment;
 import org.smartregister.cursoradapter.RecyclerViewPaginatedAdapter;
 import org.smartregister.domain.FetchStatus;
 import org.smartregister.job.SyncServiceJob;
+import org.smartregister.job.SyncSettingsServiceJob;
 import org.smartregister.provider.SmartRegisterClientsProvider;
 import org.smartregister.receiver.SyncStatusBroadcastReceiver;
 import org.smartregister.view.activity.SecuredNativeSmartRegisterActivity;
