+++ opensrp-anc/src/main/java/org/smartregister/anc/interactor/LoginInteractor.java
@@ -13,8 +13,6 @@ import org.smartregister.anc.application.AncApplication;
 import org.smartregister.anc.contract.LoginContract;
 import org.smartregister.anc.helper.CharacteristicsHelper;
 import org.smartregister.anc.job.ImageUploadServiceJob;
-import org.smartregister.anc.job.PullUniqueIdsServiceJob;
-import org.smartregister.anc.job.SyncSettingsServiceJob;
 import org.smartregister.anc.job.ViewConfigurationsServiceJob;
 import org.smartregister.anc.task.RemoteLoginTask;
 import org.smartregister.anc.util.Constants;
@@ -22,7 +20,9 @@ import org.smartregister.anc.util.NetworkUtils;
 import org.smartregister.domain.LoginResponse;
 import org.smartregister.domain.TimeStatus;
 import org.smartregister.event.Listener;
+import org.smartregister.job.PullUniqueIdsServiceJob;
 import org.smartregister.job.SyncServiceJob;
+import org.smartregister.job.SyncSettingsServiceJob;
 import org.smartregister.repository.AllSharedPreferences;
 import org.smartregister.service.UserService;
